@@ -32,12 +32,13 @@
     "@aws-sdk/credential-providers": "3.750.0",
     "@usebruno/common": "0.1.0",
     "@usebruno/converters": "^0.1.0",
+    "@usebruno/filestore": "^0.1.0",
     "@usebruno/js": "0.12.0",
     "@usebruno/lang": "0.12.0",
     "@usebruno/node-machine-id": "^2.0.0",
+    "@usebruno/requests": "^0.1.0",
     "@usebruno/schema": "0.7.0",
     "@usebruno/vm2": "^3.9.13",
-    "@usebruno/requests": "^0.1.0",
     "about-window": "^1.15.2",
     "aws4-axios": "^3.3.0",
     "axios": "^1.8.3",
@@ -64,7 +65,6 @@
     "nanoid": "3.3.8",
     "qs": "^6.11.0",
     "socks-proxy-agent": "^8.0.2",
-    "tough-cookie": "^4.1.3",
     "uuid": "^9.0.0",
     "yup": "^0.32.11"
   },
@@ -72,7 +72,7 @@
     "dmg-license": "^1.0.11"
   },
   "devDependencies": {
-    "electron": "33.2.1",
+    "electron": "~37.2.6",
     "electron-builder": "25.1.8",
     "electron-devtools-installer": "^4.0.0"
   }
