@@ -1,6 +1,6 @@
 import React from 'react';
 import Modal from 'components/Modal';
-import { IconDownload } from '@tabler/icons';
+import { IconDownload, IconLoader2 } from '@tabler/icons';
 import StyledWrapper from './StyledWrapper';
 import Bruno from 'components/Bruno';
 import exportBrunoCollection from 'utils/collections/export';
@@ -8,10 +8,12 @@ import exportPostmanCollection from 'utils/exporters/postman-collection';
 import { cloneDeep } from 'lodash';
 import { transformCollectionToSaveToExportAsFile } from 'utils/collections/index';
 import { useSelector } from 'react-redux';
-import { findCollectionByUid } from 'utils/collections/index';
+import { findCollectionByUid, areItemsLoading } from 'utils/collections/index';
 
 const ShareCollection = ({ onClose, collectionUid }) => {
   const collection = useSelector(state => findCollectionByUid(state.collections.collections, collectionUid));
+  const isCollectionLoading = areItemsLoading(collection);
+  
   const handleExportBrunoCollection = () => {
     const collectionCopy = cloneDeep(collection);
     exportBrunoCollection(transformCollectionToSaveToExportAsFile(collectionCopy));
@@ -35,23 +37,49 @@ const ShareCollection = ({ onClose, collectionUid }) => {
     >
       <StyledWrapper className="flex flex-col h-full w-[500px]">
           <div className="space-y-2"> 
-            <div className="flex border border-gray-200 dark:border-gray-600 hover:bg-gray-100 dark:hover:bg-gray-500/10 items-center p-3 rounded-lg transition-colors cursor-pointer" onClick={handleExportBrunoCollection}>
+            <div 
+              className={`flex border border-gray-200 dark:border-gray-600 items-center p-3 rounded-lg transition-colors ${
+                isCollectionLoading 
+                  ? 'opacity-50 cursor-not-allowed' 
+                  : 'hover:bg-gray-100 dark:hover:bg-gray-500/10 cursor-pointer'
+              }`}
+              onClick={isCollectionLoading ? undefined : handleExportBrunoCollection}
+            >
               <div className="mr-3 p-1 rounded-full">
-                <Bruno width={28} />
+                {isCollectionLoading ? (
+                  <IconLoader2 size={28} className="animate-spin" />
+                ) : (
+                  <Bruno width={28} />
+                )}
               </div>
               <div className="flex-1">
                 <div className="font-medium">Bruno Collection</div>
-                <div className="text-xs">Export in Bruno format</div>
+                <div className="text-xs">
+                  {isCollectionLoading ? 'Loading collection...' : 'Export in Bruno format'}
+                </div>
               </div>
             </div>
             
-            <div className="flex border border-gray-200 dark:border-gray-600 hover:bg-gray-100 dark:hover:bg-gray-500/10 items-center p-3 rounded-lg transition-colors cursor-pointer" onClick={handleExportPostmanCollection}>
+            <div 
+              className={`flex border border-gray-200 dark:border-gray-600 items-center p-3 rounded-lg transition-colors ${
+                isCollectionLoading 
+                  ? 'opacity-50 cursor-not-allowed' 
+                  : 'hover:bg-gray-100 dark:hover:bg-gray-500/10 cursor-pointer'
+              }`}
+              onClick={isCollectionLoading ? undefined : handleExportPostmanCollection}
+            >
               <div className="mr-3 p-1 rounded-full">
-                <IconDownload size={28} strokeWidth={1} className="" />
+                {isCollectionLoading ? (
+                  <IconLoader2 size={28} className="animate-spin" />
+                ) : (
+                  <IconDownload size={28} strokeWidth={1} className="" />
+                )}
               </div>
               <div className="flex-1">
                 <div className="font-medium">Postman Collection</div>
-                <div className="text-xs">Export in Postman format</div>
+                <div className="text-xs">
+                  {isCollectionLoading ? 'Loading collection...' : 'Export in Postman format'}
+                </div>
               </div>
             </div>
           </div>
