@@ -6,7 +6,7 @@ const StyledWrapper = styled.div`
     display: grid;
     overflow-y: hidden;
     overflow-x: auto;
-    padding: 0 1px;
+    padding: 0 1.5px;
 
     // for icon hover
     position: inherit;
