@@ -27,22 +27,25 @@ taskMiddleware.startListening({
     each(openRequestTasks, (task) => {
       if (collectionUid === task.collectionUid) {
         const collection = findCollectionByUid(state.collections.collections, collectionUid);
-        const item = findItemInCollectionByPathname(collection, task.itemPathname);
-        if (item) {
-          listenerApi.dispatch(
-            addTab({
-              uid: item.uid,
-              collectionUid: collection.uid,
-              requestPaneTab: getDefaultRequestPaneTab(item)
-            })
-          );
-          listenerApi.dispatch(hideHomePage());
-          listenerApi.dispatch(
-            removeTaskFromQueue({
-              taskUid: task.uid
-            })
-          );
+        if (collection && collection.mountStatus === 'mounted' && !collection.isLoading) {
+          const item = findItemInCollectionByPathname(collection, task.itemPathname);
+          if (item) {
+            listenerApi.dispatch(
+              addTab({
+                uid: item.uid,
+                collectionUid: collection.uid,
+                requestPaneTab: getDefaultRequestPaneTab(item)
+              })
+            );
+            listenerApi.dispatch(hideHomePage());
+          }
         }
+
+        listenerApi.dispatch(
+          removeTaskFromQueue({
+            taskUid: task.uid
+          })
+        );
       }
     });
   }
