@@ -38,7 +38,8 @@ import {
   setCollectionSecurityConfig,
   collectionAddOauth2CredentialsByUrl,
   collectionClearOauth2CredentialsByUrl,
-  initRunRequestEvent
+  initRunRequestEvent,
+  updateRunnerConfiguration as _updateRunnerConfiguration
 } from './index';
 
 import { each } from 'lodash';
@@ -316,9 +317,9 @@ export const cancelRunnerExecution = (cancelTokenUid) => (dispatch) => {
   cancelNetworkRequest(cancelTokenUid).catch((err) => console.log(err));
 };
 
-export const runCollectionFolder = (collectionUid, folderUid, recursive, delay, tags) => (dispatch, getState) => {
+export const runCollectionFolder = (collectionUid, folderUid, recursive, delay, tags, selectedRequestUids) => (dispatch, getState) => {
   const state = getState();
-  const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;  
+  const { globalEnvironments, activeGlobalEnvironmentUid } = state.globalEnvironments;
   const collection = findCollectionByUid(state.collections.collections, collectionUid);
 
   return new Promise((resolve, reject) => {
@@ -345,6 +346,26 @@ export const runCollectionFolder = (collectionUid, folderUid, recursive, delay, 
         collectionUid: collection.uid
       })
     );
+
+    // to only include those requests in the specified order while preserving folder data
+    if (selectedRequestUids && selectedRequestUids.length > 0) {
+      const newItems = [];
+      
+      selectedRequestUids.forEach((uid, index) => {
+        const requestItem = findItemInCollection(collectionCopy, uid);
+        if (requestItem) {
+          const clonedRequest = cloneDeep(requestItem);
+          clonedRequest.seq = index + 1;
+          newItems.push(clonedRequest);
+        }
+      });
+      
+      if (folder) {
+        folder.items = newItems;
+      } else {
+        collectionCopy.items = newItems;
+      }
+    }
 
     const { ipcRenderer } = window;
     ipcRenderer
@@ -1064,7 +1085,7 @@ export const removeCollection = (collectionUid) => (dispatch, getState) => {
     }
     const { ipcRenderer } = window;
     ipcRenderer
-      .invoke('renderer:remove-collection', collection.pathname)
+      .invoke('renderer:remove-collection', collection.pathname, collectionUid)
       .then(() => {
         dispatch(closeAllCollectionTabs({ collectionUid }));
       })
@@ -1373,3 +1394,11 @@ export const mountCollection = ({ collectionUid, collectionPathname, brunoConfig
       ipcRenderer.invoke('renderer:show-in-folder', collectionPath).then(resolve).catch(reject);
     });
   };
+
+export const updateRunnerConfiguration = (collectionUid, selectedRequestItems, requestItemsOrder) => (dispatch) => {
+  dispatch(_updateRunnerConfiguration({
+    collectionUid,
+    selectedRequestItems,
+    requestItemsOrder
+  }));
+};
