@@ -37,6 +37,7 @@ const STATIC_API_HINTS = {
     'res.headers',
     'res.body',
     'res.responseTime',
+    'res.url',
     'res.getStatus()',
     'res.getStatusText()',
     'res.getHeader(name)',
@@ -48,6 +49,7 @@ const STATIC_API_HINTS = {
     'res.getSize().header',
     'res.getSize().body',
     'res.getSize().total',
+    'res.getUrl()'
   ],
   bru: [
     'bru',
@@ -78,7 +80,17 @@ const STATIC_API_HINTS = {
     'bru.runner.setNextRequest(requestName)',
     'bru.runner.skipRequest()',
     'bru.runner.stopExecution()',
-    'bru.interpolate(str)'
+    'bru.interpolate(str)',
+    'bru.cookies',
+    'bru.cookies.jar()',
+    'bru.cookies.jar().getCookie(url, name, callback)',
+    'bru.cookies.jar().getCookies(url, callback)',
+    'bru.cookies.jar().setCookie(url, name, value, callback)',
+    'bru.cookies.jar().setCookie(url, cookieObject, callback)',
+    'bru.cookies.jar().setCookies(url, cookiesArray, callback)',
+    'bru.cookies.jar().clear(callback)',
+    'bru.cookies.jar().deleteCookies(url, callback)',
+    'bru.cookies.jar().deleteCookie(url, name, callback)',
   ]
 };
 
