@@ -20,7 +20,7 @@ const { addAwsV4Interceptor, resolveAwsV4Credentials } = require('./awsv4auth-he
 const { shouldUseProxy, PatchedHttpsProxyAgent, getSystemProxyEnvVariables } = require('../utils/proxy-util');
 const path = require('path');
 const { parseDataFromResponse } = require('../utils/common');
-const { getCookieStringForUrl, saveCookies, shouldUseCookies } = require('../utils/cookies');
+const { getCookieStringForUrl, saveCookies } = require('../utils/cookies');
 const { createFormData } = require('../utils/form-data');
 const { getOAuth2Token } = require('./oauth2');
 const protocolRegex = /^([-+\w]{1,25})(:?\/\/|:)/;
@@ -287,6 +287,10 @@ const runSingleRequest = async function (
               https_proxy,
               Object.keys(httpsAgentRequestFields).length > 0 ? { ...httpsAgentRequestFields } : undefined
             );
+          } else {
+            request.httpsAgent = new https.Agent({
+              ...httpsAgentRequestFields
+            });
           }
         } catch (error) {
           throw new Error('Invalid system https_proxy');
@@ -459,6 +463,7 @@ const runSingleRequest = async function (
             statusText: null,
             headers: null,
             data: null,
+            url: null,
             responseTime: 0
           },
           error: err?.message || err?.errors?.map(e => e?.message)?.at(0) || err?.code || 'Request Failed!',
@@ -598,6 +603,7 @@ const runSingleRequest = async function (
         statusText: response.statusText,
         headers: response.headers,
         data: response.data,
+        url: response.request ? response.request.protocol + '//' + response.request.host + response.request.path : null,
         responseTime
       },
       error: null,
@@ -626,6 +632,7 @@ const runSingleRequest = async function (
         statusText: null,
         headers: null,
         data: null,
+        url: null,
         responseTime: 0
       },
       status: 'error',
