@@ -11,22 +11,35 @@ describe('parseBruFileMeta', () => {
     const result = parseBruFileMeta(data);
 
     expect(result).toEqual({
-      meta: {
-        name: '0.2_mb',
-        type: 'http',
-        seq: 1,
-      },
+      type: 'http-request',
+      name: '0.2_mb',
+      seq: 1,
+      settings: {},
+      tags: [],
+      request: {
+        method: '',
+        url: '',
+        params: [],
+        headers: [],
+        auth: { mode: 'none' },
+        body: { mode: 'none' },
+        script: {},
+        vars: {},
+        assertions: [],
+        tests: '',
+        docs: ''
+      }
     });
   });
 
-  test('returns undefined for missing meta block', () => {
+  test('returns null for missing meta block', () => {
     const data = `someOtherBlock {
       key: value
     }`;
 
     const result = parseBruFileMeta(data);
 
-    expect(result).toBeUndefined();
+    expect(result).toBeNull();
   });
 
   test('handles empty meta block gracefully', () => {
@@ -34,7 +47,26 @@ describe('parseBruFileMeta', () => {
 
     const result = parseBruFileMeta(data);
 
-    expect(result).toEqual({ meta: {} });
+    expect(result).toEqual({
+      type: 'http-request',
+      name: undefined,
+      seq: 1,
+      settings: {},
+      tags: [],
+      request: {
+        method: '',
+        url: '',
+        params: [],
+        headers: [],
+        auth: { mode: 'none' },
+        body: { mode: 'none' },
+        script: {},
+        vars: {},
+        assertions: [],
+        tests: '',
+        docs: ''
+      }
+    });
   });
 
   test('ignores invalid lines in meta block', () => {
@@ -47,10 +79,24 @@ describe('parseBruFileMeta', () => {
     const result = parseBruFileMeta(data);
 
     expect(result).toEqual({
-      meta: {
-        name: '0.2_mb',
-        seq: 1,
-      },
+      type: 'http-request',
+      name: '0.2_mb',
+      seq: 1,
+      settings: {},
+      tags: [],
+      request: {
+        method: '',
+        url: '',
+        params: [],
+        headers: [],
+        auth: { mode: 'none' },
+        body: { mode: 'none' },
+        script: {},
+        vars: {},
+        assertions: [],
+        tests: '',
+        docs: ''
+      }
     });
   });
 
@@ -59,7 +105,7 @@ describe('parseBruFileMeta', () => {
 
     const result = parseBruFileMeta(data);
 
-    expect(result).toBeUndefined();
+    expect(result).toBeNull();
   });
 
   test('handles missing colon gracefully', () => {
@@ -71,9 +117,24 @@ describe('parseBruFileMeta', () => {
     const result = parseBruFileMeta(data);
 
     expect(result).toEqual({
-      meta: {
-        seq: 1,
-      },
+      type: 'http-request',
+      name: undefined,
+      seq: 1,
+      settings: {},
+      tags: [],
+      request: {
+        method: '',
+        url: '',
+        params: [],
+        headers: [],
+        auth: { mode: 'none' },
+        body: { mode: 'none' },
+        script: {},
+        vars: {},
+        assertions: [],
+        tests: '',
+        docs: ''
+      }
     });
   });
 
@@ -82,16 +143,30 @@ describe('parseBruFileMeta', () => {
       numValue: 1234
       floatValue: 12.34
       strValue: some_text
+      seq: 5
     }`;
 
     const result = parseBruFileMeta(data);
 
     expect(result).toEqual({
-      meta: {
-        numValue: 1234,
-        floatValue: 12.34,
-        strValue: 'some_text',
-      },
+      type: 'http-request',
+      name: undefined,
+      seq: 5,
+      settings: {},
+      tags: [],
+      request: {
+        method: '',
+        url: '',
+        params: [],
+        headers: [],
+        auth: { mode: 'none' },
+        body: { mode: 'none' },
+        script: {},
+        vars: {},
+        assertions: [],
+        tests: '',
+        docs: ''
+      }
     });
   });
 
@@ -104,7 +179,7 @@ describe('parseBruFileMeta', () => {
 
     const result = parseBruFileMeta(data);
 
-    expect(result).toBeUndefined();
+    expect(result).toBeNull();
   });
 
   test('handles syntax error in meta block 2', () => {
@@ -116,6 +191,98 @@ describe('parseBruFileMeta', () => {
 
     const result = parseBruFileMeta(data);
 
-    expect(result).toBeUndefined();
+    expect(result).toBeNull();
+  });
+
+  test('handles graphql type correctly', () => {
+    const data = `meta {
+      name: graphql_query
+      type: graphql
+      seq: 2
+    }`;
+
+    const result = parseBruFileMeta(data);
+
+    expect(result).toEqual({
+      type: 'graphql-request',
+      name: 'graphql_query',
+      seq: 2,
+      settings: {},
+      tags: [],
+      request: {
+        method: '',
+        url: '',
+        params: [],
+        headers: [],
+        auth: { mode: 'none' },
+        body: { mode: 'none' },
+        script: {},
+        vars: {},
+        assertions: [],
+        tests: '',
+        docs: ''
+      }
+    });
+  });
+
+  test('handles unknown type correctly', () => {
+    const data = `meta {
+      name: unknown_request
+      type: unknown
+      seq: 3
+    }`;
+
+    const result = parseBruFileMeta(data);
+
+    expect(result).toEqual({
+      type: 'http-request',
+      name: 'unknown_request',
+      seq: 3,
+      settings: {},
+      tags: [],
+      request: {
+        method: '',
+        url: '',
+        params: [],
+        headers: [],
+        auth: { mode: 'none' },
+        body: { mode: 'none' },
+        script: {},
+        vars: {},
+        assertions: [],
+        tests: '',
+        docs: ''
+      }
+    });
+  });
+
+  test('handles missing seq gracefully', () => {
+    const data = `meta {
+      name: no_seq_request
+      type: http
+    }`;
+
+    const result = parseBruFileMeta(data);
+
+    expect(result).toEqual({
+      type: 'http-request',
+      name: 'no_seq_request',
+      seq: 1, // Default fallback
+      settings: {},
+      tags: [],
+      request: {
+        method: '',
+        url: '',
+        params: [],
+        headers: [],
+        auth: { mode: 'none' },
+        body: { mode: 'none' },
+        script: {},
+        vars: {},
+        assertions: [],
+        tests: '',
+        docs: ''
+      }
+    });
   });
 });
