@@ -258,6 +258,136 @@ const addBruShimToContext = (vm, bru) => {
   });
   sleep.consume((handle) => vm.setProp(bruObject, 'sleep', handle));
 
+  let bruCookiesObject = vm.newObject();
+
+  const _jarFn = vm.newFunction('_jar', () => {
+    const nativeJar = bru.cookies.jar();
+    const jarObj = vm.newObject();
+
+    const _getCookieFn = vm.newFunction('_getCookie', (url, cookieName) => {
+      const promise = vm.newPromise();
+      nativeJar.getCookie(vm.dump(url), vm.dump(cookieName), (err, cookie) => {
+        if (err) {
+          promise.reject(marshallToVm(cleanJson(err), vm));
+        } else {
+          promise.resolve(marshallToVm(cleanCircularJson(cookie), vm));
+        }
+      });
+      promise.settled.then(vm.runtime.executePendingJobs);
+      return promise.handle;
+    });
+    _getCookieFn.consume((handle) => vm.setProp(jarObj, '_getCookie', handle));
+
+    const _getCookiesFn = vm.newFunction('_getCookies', (url) => {
+      const promise = vm.newPromise();
+      nativeJar.getCookies(vm.dump(url), (err, cookies) => {
+        if (err) {
+          promise.reject(marshallToVm(cleanJson(err), vm));
+        } else {
+          promise.resolve(marshallToVm(cleanCircularJson(cookies), vm));
+        }
+      });
+      promise.settled.then(vm.runtime.executePendingJobs);
+      return promise.handle;
+    });
+    _getCookiesFn.consume((handle) => vm.setProp(jarObj, '_getCookies', handle));
+
+    const _setCookieFn = vm.newFunction('_setCookie', (url, nameOrCookieObj, value) => {
+      const promise = vm.newPromise();
+      const dumpedUrl = vm.dump(url);
+      const dumpedNameOrObj = vm.dump(nameOrCookieObj);
+      
+      // Check if the second argument is an object (cookie object case)
+      if (typeof dumpedNameOrObj === 'object' && dumpedNameOrObj !== null) {
+        // Cookie object case: setCookie(url, cookieObject, callback)
+        nativeJar.setCookie(dumpedUrl, dumpedNameOrObj, (err) => {
+          if (err) {
+            promise.reject(marshallToVm(cleanJson(err), vm));
+          } else {
+            promise.resolve(vm.undefined);
+          }
+        });
+      } else {
+        // Name/value case: setCookie(url, name, value, callback)
+        const dumpedValue = value ? vm.dump(value) : '';
+        nativeJar.setCookie(dumpedUrl, dumpedNameOrObj, dumpedValue, (err) => {
+          if (err) {
+            promise.reject(marshallToVm(cleanJson(err), vm));
+          } else {
+            promise.resolve(vm.undefined);
+          }
+        });
+      }
+      
+      promise.settled.then(vm.runtime.executePendingJobs);
+      return promise.handle;
+    });
+    _setCookieFn.consume((handle) => vm.setProp(jarObj, '_setCookie', handle));
+
+    const _setCookiesFn = vm.newFunction('_setCookies', (url, cookiesArray) => {
+      const promise = vm.newPromise();
+      
+      nativeJar.setCookies(vm.dump(url), vm.dump(cookiesArray), (err) => {
+        if (err) {
+          promise.reject(marshallToVm(cleanJson(err), vm));
+        } else {
+          promise.resolve(vm.undefined);
+        }
+      });
+      promise.settled.then(vm.runtime.executePendingJobs);
+      return promise.handle;
+    });
+    _setCookiesFn.consume((handle) => vm.setProp(jarObj, '_setCookies', handle));
+
+    const _clearFn = vm.newFunction('_clear', () => {
+      const promise = vm.newPromise();
+      nativeJar.clear((err) => {
+        if (err) {
+          promise.reject(marshallToVm(cleanJson(err), vm));
+        } else {
+          promise.resolve(vm.undefined);
+        }
+      });
+      promise.settled.then(vm.runtime.executePendingJobs);
+      return promise.handle;
+    });
+    _clearFn.consume((handle) => vm.setProp(jarObj, '_clear', handle));
+
+    const _deleteCookiesFn = vm.newFunction('_deleteCookies', (url) => {
+      const promise = vm.newPromise();
+      nativeJar.deleteCookies(vm.dump(url), (err) => {
+        if (err) {
+          promise.reject(marshallToVm(cleanJson(err), vm));
+        } else {
+          promise.resolve(vm.undefined);
+        }
+      });
+      promise.settled.then(vm.runtime.executePendingJobs);
+      return promise.handle;
+    });
+    _deleteCookiesFn.consume((handle) => vm.setProp(jarObj, '_deleteCookies', handle));
+
+    const _deleteCookieFn = vm.newFunction('_deleteCookie', (url, cookieName) => {
+      const promise = vm.newPromise();
+      nativeJar.deleteCookie(vm.dump(url), vm.dump(cookieName), (err) => {
+        if (err) {
+          promise.reject(marshallToVm(cleanJson(err), vm));
+        } else {
+          promise.resolve(vm.undefined);
+        }
+      });
+      promise.settled.then(vm.runtime.executePendingJobs);
+      return promise.handle;
+    });
+    _deleteCookieFn.consume((handle) => vm.setProp(jarObj, '_deleteCookie', handle));
+
+    return jarObj;
+  });
+  _jarFn.consume((handle) => vm.setProp(bruCookiesObject, '_jar', handle));
+
+  vm.setProp(bruObject, 'cookies', bruCookiesObject);
+  bruCookiesObject.dispose();
+
   vm.setProp(bruObject, 'runner', bruRunnerObject);
   vm.setProp(vm.global, 'bru', bruObject);
   bruObject.dispose();
@@ -282,7 +412,41 @@ const addBruShimToContext = (vm, bru) => {
           return Promise.reject(err);
         }
       }
-    }
+    };
+
+    globalThis.bru.cookies.jar = () => {
+      const _jar = globalThis.bru.cookies._jar();
+
+      const callWithCallback = async (promiseFn, callback) => {
+        if (!callback) return await promiseFn();
+        try {
+          const result = await promiseFn();
+          try { await callback(null, result); } catch(cbErr) { return Promise.reject(cbErr); }
+        } catch(err) {
+          try { await callback(err, null); } catch(cbErr) { return Promise.reject(cbErr); }
+        }
+      };
+
+      return {
+        getCookie: (url, name, cb) => callWithCallback(() => _jar._getCookie(url, name), cb),
+        getCookies: (url, cb) => callWithCallback(() => _jar._getCookies(url), cb),
+        setCookie: (url, nameOrCookieObj, valueOrCallback, maybeCallback) => {
+          if (typeof nameOrCookieObj === 'object' && nameOrCookieObj !== null) {
+            const callback = typeof valueOrCallback === 'function' ? valueOrCallback : undefined;
+            return callWithCallback(() => _jar._setCookie(url, nameOrCookieObj), callback);
+          } else {
+            const value = typeof valueOrCallback === 'string' ? valueOrCallback : '';
+            const callback = typeof maybeCallback === 'function' ? maybeCallback : 
+                           (typeof valueOrCallback === 'function' ? valueOrCallback : undefined);
+            return callWithCallback(() => _jar._setCookie(url, nameOrCookieObj, value), callback);
+          }
+        },
+        setCookies: (url, cookiesArray, cb) => callWithCallback(() => _jar._setCookies(url, cookiesArray), cb),
+        clear: (cb) => callWithCallback(() => _jar._clear(), cb),
+        deleteCookies: (url, cb) => callWithCallback(() => _jar._deleteCookies(url), cb),
+        deleteCookie: (url, name, cb) => callWithCallback(() => _jar._deleteCookie(url, name), cb)
+      };
+    };
   `);
 };
 
